requirements.txt
# main
Django==4.2.1
djangorestframework==3.14.0
gunicorn==20.1.0

# debug
django-debug-toolbar==4.1.0

# token
djangorestframework-simplejwt==5.2.2

# Security
django-environ==0.10.0
django-cors-headers==4.0.0

# For OpenAPI
drf-spectacular==0.26.2
drf-standardized-errors==0.12.5


# db
psycopg2==2.9.6
django-filter==23.2

# extra
Pillow==9.5.0
django-cleanup==8.0.0

# linter
ruff==0.0.270

#reset-password
django-rest-passwordreset==1.3.0

#celery
celery==5.3.0
redis==4.5.5
#django-celery-beat==2.5.0

#open graph
topicaxis-opengraph==0.5